refactor: 將 DTO 驗證搬移出去
