+++ src/auth/auth.controller.spec.ts
@@ -1,10 +1,4 @@
-import {
-  ArgumentMetadata,
-  BadRequestException,
-  ConflictException,
-  HttpStatus,
-  ValidationPipe,
-} from "@nestjs/common";
+import { ConflictException } from "@nestjs/common";
 import { Test, TestingModule } from "@nestjs/testing";
 import { getRepositoryToken, TypeOrmModule } from "@nestjs/typeorm";
 import { dataSourceJest } from "src/config/data-source";
@@ -21,13 +15,6 @@ describe("AuthController", () => {
   let authController: AuthController;
   let authService: AuthService;
   let userRepository: Repository<UserEntity>;
-  const target: ValidationPipe = new ValidationPipe({
-    errorHttpStatusCode: HttpStatus.BAD_REQUEST,
-    stopAtFirstError: false,
-    disableErrorMessages: false,
-    transform: true,
-    whitelist: true,
-  });
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [TypeOrmModule.forRoot(dataSourceJest)],
@@ -48,137 +35,6 @@ describe("AuthController", () => {
       getRepositoryToken(UserEntity),
     );
   });
-  describe("createUser - DTO", () => {
-    it("應該會發生 email 欄位未填驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "",
-        name: "displayname",
-        account: "account",
-        password: "Password@123",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: ["email 為必填欄位。", "email 必須是信箱格式。"],
-          error: "Bad Request",
-        });
-      });
-    });
-    it("應該會發生 email 欄位格式驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "jhon",
-        name: "displayname",
-        account: "account",
-        password: "Password@123",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: ["email 必須是信箱格式。"],
-          error: "Bad Request",
-        });
-      });
-    });
-    it("應該會發生 name 欄位未填驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "jhon@gmail.com",
-        name: "",
-        account: "account",
-        password: "Password@123",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: ["name 為必填欄位。"],
-          error: "Bad Request",
-        });
-      });
-    });
-    it("應該會發生 account 欄位未填驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "jhon@gmail.com",
-        name: "displayname",
-        account: "",
-        password: "Password@123",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: ["account 為必填欄位。"],
-          error: "Bad Request",
-        });
-      });
-    });
-    it("應該會發生 password 欄位未填驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "jhon@gmail.com",
-        name: "displayname",
-        account: "account",
-        password: "",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: [
-            "password 必須長度大於等於8個字。",
-            "password 為必填欄位。",
-          ],
-          error: "Bad Request",
-        });
-      });
-    });
-    it("應該會發生 password 欄位長度驗證失敗", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "jhon@gmail.com",
-        name: "displayname",
-        account: "account",
-        password: "123",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: ["password 必須長度大於等於8個字。"],
-          error: "Bad Request",
-        });
-      });
-    });
-  });
   describe("create", () => {
     it("應該會創建一個使用者，並返回 201 狀態碼", async () => {
       const createUserDto: CreateUserDto = {
@@ -194,34 +50,6 @@ describe("AuthController", () => {
       jest.spyOn(authService, "register").mockResolvedValue(expectedResponse);
       const result = await authController.register(createUserDto);
       expect(result).toEqual(expectedResponse);
-    });
-    it("應該會發生資料驗證失敗，並返回 400 狀態碼", async () => {
-      const createUserDto: CreateUserDto = {
-        email: "",
-        name: "",
-        account: "",
-        password: "",
-      };
-      const metadata: ArgumentMetadata = {
-        type: "body",
-        metatype: CreateUserDto,
-        data: "@Body()",
-      };
-      await target.transform(createUserDto, metadata).catch(error => {
-        expect(error).toBeInstanceOf(BadRequestException);
-        expect(error.response).toEqual({
-          statusCode: 400,
-          message: [
-            "email 為必填欄位。",
-            "email 必須是信箱格式。",
-            "name 為必填欄位。",
-            "account 為必填欄位。",
-            "password 必須長度大於等於8個字。",
-            "password 為必填欄位。",
-          ],
-          error: "Bad Request",
-        });
-      });
     });
     it("應該會發生資料使用者重覆，並返回 409 狀態碼", async () => {
       const createUserDto1: CreateUserDto = {
